nexus-daily/ingestion/requirements.txt
requests
beautifulsoup4
feedparser
openai
python-dotenv
prisma
